requirements.txt
Pillow
pyyaml
python-periphery
flask
paho-mqtt
cryptography
